Refactor filesystem.py to python package
